pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "subcellspace"
version = "0.1.0"
description = "A modular subcellular spatial transcriptomics analysis platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anndata>=0.12",
  "fastapi>=0.115",
  "numpy>=2.0",
  "pandas>=2.2",
  "pydantic>=2.8",
  "pyyaml>=6.0",
  "scanpy>=1.11",
  "scikit-learn>=1.5",
  "scipy>=1.13",
  "spatialdata>=0.3",
  "squidpy>=1.7",
  "geopandas>=1.0",
  "shapely>=2.0",
  "uvicorn[standard]>=0.30",
  "igraph>=0.11",
  "leidenalg>=0.10",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=4.0", "ruff>=0.6", "mypy>=1.11", "httpx>=0.28.1"]

# Denoise backends
sparc = []  # installed via -e tools/spARC/

# Segmentation backends
cellpose = ["cellpose>=3.0"]  # requires external microscopy image (DAPI) for segmentation pipeline step
baysor = []  # NOT available: requires Julia runtime; not tracked in Python toolchain

# Spatial domain backends
graphst = ["pot"]
stagate = []  # needs tensorflow
spagcn = []   # needs compilation

# Spatial analysis backends (Phase 6)
scfates = []  # installed via -e tools/scFates/ (tree inference + pseudotime)
snakemake = ["snakemake>=9.0"]  # workflow engine for patchify scheduling

# Subcellular spatial domain backends
phenograph = []  # installed via -e tools/PhenoGraph/
hdbscan = ["hdbscan>=0.8.42"]

# Analysis backends
scvi = ["scvi-tools"]

# Annotation backends
celltypist = []  # installed via -e tools/celltypist/

# All optional tools combined (Python-installable only; excludes baysor)
all = ["subcellspace[dev,sparc,cellpose,graphst,phenograph,hdbscan,scvi,celltypist]"]


[project.scripts]
subcellspace = "src.cli:main"
subcellspace-api = "src.api_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
]
ignore = ["E501"]  # line-length is handled by ruff formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
disable_error_code = [
    "type-arg",      # numpy/scipy/pandas stubs not fully aligned
    "var-annotated", # some untyped class vars ok in dataclasses
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "tools/*", "scripts/*"]

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "raise NotImplementedError",
    "def __repr__",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]

